pyproject.toml
[project]
name = "voice-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "cohere>=5.0.0",
    "fastapi>=0.117.1",
    "mypy>=1.18.2",
    "pydantic-settings>=2.11.0",
    "psycopg2-binary>=2.9.9",
    "requests>=2.31.0",
    "ruff>=0.13.1",
    "sqlmodel>=0.0.17",
    "tenacity>=9.1.2",
    "uvicorn>=0.37.0",
    "black>=25.9.0",
    "isort>=6.0.1",
    "alembic>=1.16.5",
]
